LPD-22831 Rename base and associated classes
