+++ modules/apps/list-type/list-type-service/src/main/java/com/liferay/list/type/internal/upgrade/registry/ListTypeServiceUpgradeStepRegistrator.java
@@ -6,8 +6,8 @@
 package com.liferay.list.type.internal.upgrade.registry;
 
 import com.liferay.list.type.internal.upgrade.v1_3_0.ListTypeDefinitionUpgradeProcess;
-import com.liferay.list.type.internal.upgrade.v1_3_2.ListTypeDefinitionUserIdUpgradeProcess;
-import com.liferay.list.type.internal.upgrade.v1_3_2.ListTypeEntryUserIdUpgradeProcess;
+import com.liferay.list.type.internal.upgrade.v1_3_2.ListTypeDefinitionStaleUserIdUpgradeProcess;
+import com.liferay.list.type.internal.upgrade.v1_3_2.ListTypeEntryStaleUserIdUpgradeProcess;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.model.ListTypeConstants;
 import com.liferay.portal.kernel.service.UserLocalService;
@@ -76,8 +76,8 @@ public class ListTypeServiceUpgradeStepRegistrator
 
 		registry.register(
 			"1.3.1", "1.3.2",
-			new ListTypeDefinitionUserIdUpgradeProcess(_userLocalService),
-			new ListTypeEntryUserIdUpgradeProcess(_userLocalService));
+			new ListTypeDefinitionStaleUserIdUpgradeProcess(_userLocalService),
+			new ListTypeEntryStaleUserIdUpgradeProcess(_userLocalService));
 	}
 
 	@Reference
