+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/upgrade/v10_0_2/ObjectFieldStaleUserIdUpgradeProcess.java
@@ -6,14 +6,17 @@
 package com.liferay.object.internal.upgrade.v10_0_2;
 
 import com.liferay.portal.kernel.service.UserLocalService;
-import com.liferay.portal.kernel.upgrade.BaseUserIdUpgradeProcess;
+import com.liferay.portal.kernel.upgrade.BaseStaleUserIdUpgradeProcess;
 
 /**
  * @author Igor Costa
  */
-public class ObjectFieldUserIdUpgradeProcess extends BaseUserIdUpgradeProcess {
+public class ObjectFieldStaleUserIdUpgradeProcess
+	extends BaseStaleUserIdUpgradeProcess {
 
-	public ObjectFieldUserIdUpgradeProcess(UserLocalService userLocalService) {
+	public ObjectFieldStaleUserIdUpgradeProcess(
+		UserLocalService userLocalService) {
+
 		super(userLocalService);
 	}
