+++ modules/apps/object/object-test/src/testIntegration/java/com/liferay/object/internal/upgrade/v10_0_2/test/ObjectDefinitionStaleUserIdUpgradeProcessTest.java
@@ -3,27 +3,23 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.list.type.internal.upgrade.v1_3_2.test;
+package com.liferay.object.internal.upgrade.v10_0_2.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.list.type.entry.util.ListTypeEntryUtil;
-import com.liferay.list.type.model.ListTypeDefinition;
-import com.liferay.list.type.service.ListTypeDefinitionLocalService;
+import com.liferay.object.model.ObjectDefinition;
+import com.liferay.object.service.ObjectDefinitionLocalService;
+import com.liferay.object.test.util.ObjectDefinitionTestUtil;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.kernel.test.util.UserTestUtil;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
 import com.liferay.portal.upgrade.test.util.UpgradeTestUtil;
-
-import java.util.Collections;
 
 import org.junit.After;
 import org.junit.Assert;
@@ -37,7 +33,7 @@ import org.junit.runner.RunWith;
  * @author Igor Costa
  */
 @RunWith(Arquillian.class)
-public class ListTypeDefinitionUserIdUpgradeProcessTest {
+public class ObjectDefinitionStaleUserIdUpgradeProcessTest {
 
 	@ClassRule
 	@Rule
@@ -50,21 +46,17 @@ public class ListTypeDefinitionUserIdUpgradeProcessTest {
 	public static void setUpClass() throws Exception {
 		_user = UserTestUtil.addUser(TestPropsValues.getCompanyId());
 
-		_listTypeDefinition =
-			_listTypeDefinitionLocalService.addListTypeDefinition(
-				null, _user.getUserId(),
-				Collections.singletonMap(
-					LocaleUtil.US, RandomTestUtil.randomString()),
-				false,
-				Collections.singletonList(
-					ListTypeEntryUtil.createListTypeEntry(
-						RandomTestUtil.randomString())));
+		_objectDefinition = ObjectDefinitionTestUtil.addCustomObjectDefinition(
+			ObjectDefinitionTestUtil.getRandomName(), _user.getUserId());
+
+		_objectDefinition =
+			_objectDefinitionLocalService.publishCustomObjectDefinition(
+				_user.getUserId(), _objectDefinition.getObjectDefinitionId());
 	}
 
 	@After
 	public void tearDown() throws Exception {
-		_listTypeDefinitionLocalService.deleteListTypeDefinition(
-			_listTypeDefinition);
+		_objectDefinitionLocalService.deleteObjectDefinition(_objectDefinition);
 	}
 
 	@Test
@@ -75,37 +67,33 @@ public class ListTypeDefinitionUserIdUpgradeProcessTest {
 
 		Assert.assertNull(_userLocalService.fetchUser(userId));
 
-		Assert.assertEquals(userId, _listTypeDefinition.getUserId());
+		Assert.assertEquals(userId, _objectDefinition.getUserId());
 
 		UpgradeProcess upgradeProcess = UpgradeTestUtil.getUpgradeStep(
 			_upgradeStepRegistrator,
-			"com.liferay.list.type.internal.upgrade.v1_3_2." +
-				"ListTypeDefinitionUserIdUpgradeProcess");
+			"com.liferay.object.internal.upgrade.v10_0_2." +
+				"ObjectDefinitionStaleUserIdUpgradeProcess");
 
 		upgradeProcess.upgrade();
 
 		User defaultServiceAccountUser =
 			_userLocalService.fetchUserByScreenName(
 				TestPropsValues.getCompanyId(), "default-service-account");
-		_listTypeDefinition =
-			_listTypeDefinitionLocalService.
-				getListTypeDefinitionByUuidAndCompanyId(
-					_listTypeDefinition.getUuid(),
-					TestPropsValues.getCompanyId());
+		_objectDefinition = _objectDefinitionLocalService.getObjectDefinition(
+			_objectDefinition.getObjectDefinitionId());
 
 		Assert.assertEquals(
 			defaultServiceAccountUser.getUserId(),
-			_listTypeDefinition.getUserId());
+			_objectDefinition.getUserId());
 	}
 
-	private static ListTypeDefinition _listTypeDefinition;
+	private static ObjectDefinition _objectDefinition;
 
 	@Inject
-	private static ListTypeDefinitionLocalService
-		_listTypeDefinitionLocalService;
+	private static ObjectDefinitionLocalService _objectDefinitionLocalService;
 
 	@Inject(
-		filter = "component.name=com.liferay.list.type.internal.upgrade.registry.ListTypeServiceUpgradeStepRegistrator"
+		filter = "component.name=com.liferay.object.internal.upgrade.registry.ObjectServiceUpgradeStepRegistrator"
 	)
 	private static UpgradeStepRegistrator _upgradeStepRegistrator;
